Parameterize bare list types in stubs
- extract_unresolved: detect bare list args, returns, and properties
- generate_stubs: parameterize list properties using element_repr
- Regenerate stubs with LLM-resolved list types

+++ stubs/12.3.6/Live/Sample/Sample.pyi
@@ -460,7 +460,7 @@ class Sample:
         ...
 
     @property
-    def slices(self) -> tuple[float, ...]:
+    def slices(self) -> tuple[int, ...]:
         """Access to the list of slice points in sample time in the sample."""
         ...
